Updates to code
I added a function that takes care of the beeping for me so I can make the code look nicer. Figured a lot of things would beep so I created a function that could do the beeping for me. Simplified the player class with ternary if statements. Trying to figure out the ammo and health stuff so a game over can be reached. At the moment the reset() function isn't working as I want it to. loop() continually spits out what is in reset() ddespite the if statement not being true.